+++ pkgs/firehose/CHANGELOG.md
@@ -1,8 +1,9 @@
-## 0.10.2-wip
+## 0.10.2
 
 - Don't check licenses of generated files in PR health workflow.
 - Add generalized ignore-per-checks to health workflow.
 - Update dart_apitool version in health workflow.
+- Print detailed info about the leaked APIs to stdout in the workflow.
 
 ## 0.10.1